Fix: CSS/JS for make:auth didn't work with MAMP
This issue occurred on Mac OS Sierra. The app.css and app.js weren't able to be found when the app was run with MAMP, but were able to able to be found when running "php artisan serve" in the command line. This makes it able for both to be found.