Remove X-UA-Compatible meta tag (#25442)
* Remove X-UA-Compatible meta tag

No reason to include this anymore. Microsoft do not support IE8, 9 and 10, and this tag is only for them. It also goes against best practices: https://stackoverflow.com/a/26348511/199700

* Remove X-UA-Compatible meta tag

No reason to include this anymore. Microsoft do not support IE8, 9 and 10, and this tag is only for them. It also goes against best practices: https://stackoverflow.com/a/26348511/199700